Create the DAO interface classes and their implementations with methods test and adjust in other classes fixes #6

+++ src/main/java/com/conexia/dao/CooksDAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface CooksDAO {
 
-    public CooksEntity  selecyById(Long idCook);
+    public CooksEntity  selecyById(Integer idCook);
 
     public List<CooksEntity> selectAll();
